refactor: 동시 실행 함수 클래스 분리
