+++ commerce-api/src/test/kotlin/com/hanghae/commerce/order/application/OrderCreateFacadeTest.kt
@@ -1,6 +1,5 @@
 package com.hanghae.commerce.order.application
 
-import com.hanghae.commerce.event.EventPublisher
 import com.hanghae.commerce.item.domain.Item
 import com.hanghae.commerce.item.domain.ItemRepository
 import com.hanghae.commerce.order.domain.OrderRepository
@@ -10,12 +9,10 @@ import com.hanghae.commerce.order.presentaion.dto.OrderCreateRequest
 import com.hanghae.commerce.order.presentaion.dto.OrderCreateResponse
 import com.hanghae.commerce.testconfiguration.EnableTestcontainers
 import com.hanghae.commerce.testconfiguration.IntegrationTest
+import com.hanghae.commerce.tools.TestConcurrentExecutor
 import org.assertj.core.api.Assertions.*
 import org.junit.jupiter.api.*
-import org.mockito.ArgumentMatchers.*
-import org.mockito.Mockito.*
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.mock.mockito.MockBean
 import java.util.concurrent.*
 
 @IntegrationTest
@@ -23,8 +20,8 @@ import java.util.concurrent.*
 @DisplayName("Given: orderCreate()")
 class OrderCreateFacadeTest {
 
-    @MockBean
-    private lateinit var eventPublisher: EventPublisher
+    @Autowired
+    private lateinit var testConcurrentExecutor: TestConcurrentExecutor
 
     @Autowired
     private lateinit var orderRepository: OrderRepository
@@ -115,7 +112,7 @@ class OrderCreateFacadeTest {
         @Test
         @DisplayName("Then: 요청 300개가 동시에 3개씩 주문하면, 남는 재고는 100개이다.")
         fun tc1() {
-            executeOrderInMultiThread(
+            testConcurrentExecutor.executeOrderInMultiThread(
                 300,
             ) {
                 sut.create(
@@ -136,7 +133,7 @@ class OrderCreateFacadeTest {
         @DisplayName("Then: 요청 200개가 동시에 5개씩 주문하면, 남는 재고는 0개이다.")
         fun tc2() {
             // when
-            executeOrderInMultiThread(
+            testConcurrentExecutor.executeOrderInMultiThread(
                 threadCount = 200,
             ) {
                 sut.create(
@@ -157,7 +154,7 @@ class OrderCreateFacadeTest {
         @DisplayName("Then: 요청 300개가 동시에 4개씩 주문하면, 요청 50개는 SoldOutException 발생한다.")
         fun tc3() {
             // when
-            val futures = executableFutures<OrderCreateResponse>(
+            val futures = testConcurrentExecutor.executableFutures<OrderCreateResponse>(
                 threadCount = 300,
             ) {
                 sut.create(
@@ -185,38 +182,6 @@ class OrderCreateFacadeTest {
             assertThat(results).hasSize(250)
             assertThat(soldOutExceptions).hasSize(50)
             assertThat(soldOutExceptions.map { it!!.message }.toList()).containsOnly("재고가 부족합니다.")
-        }
-
-        private fun <T> executableFutures(
-            threadCount: Int,
-            callable: Callable<T>,
-        ): List<Future<T>> {
-            val executorService = Executors.newFixedThreadPool(32)
-            val futures: MutableList<Future<T>> = mutableListOf()
-
-            for (i in 0 until threadCount) {
-                futures.add(executorService.submit(callable) as Future<T>)
-            }
-
-            return futures
-        }
-
-        @Throws(InterruptedException::class)
-        private fun executeOrderInMultiThread(
-            threadCount: Int,
-            runnable: Runnable,
-        ) {
-            val executorService = Executors.newFixedThreadPool(32)
-            val latch = CountDownLatch(threadCount)
-
-            for (i in 0 until threadCount) {
-                try {
-                    executorService.submit(runnable)
-                } finally {
-                    latch.countDown()
-                }
-            }
-            latch.await()
         }
     }
